Restructure HTTP code for shared use with Timeline and add add extension availability check.
