+++ packages/devtools_app/lib/src/http/_http_exception.dart
