+++ packages/devtools_app/test/flutter/network_profiler_test.dart
@@ -9,8 +9,8 @@ import 'package:devtools_app/src/network/flutter/http_request_inspector.dart';
 import 'package:devtools_app/src/network/flutter/http_request_inspector_views.dart';
 import 'package:devtools_app/src/network/flutter/network_model.dart';
 import 'package:devtools_app/src/network/flutter/network_screen.dart';
-import 'package:devtools_app/src/network/http.dart';
-import 'package:devtools_app/src/network/http_request_data.dart';
+import 'package:devtools_app/src/http/http.dart';
+import 'package:devtools_app/src/http/http_request_data.dart';
 import 'package:devtools_app/src/network/network_controller.dart';
 import 'package:devtools_app/src/service_manager.dart';
 import 'package:devtools_app/src/ui/fake_flutter/_real_flutter.dart';
