+++ packages/devtools_app/test/support/mocks.dart
@@ -204,7 +204,7 @@ class FakeVmService extends Fake implements VmServiceWrapper {
   Future<Success> clearCpuSamples(String isolateId) => Future.value(Success());
 
   @override
-  Future<HttpTimelineLoggingState> httpEnableTimelineLogging(
+  Future<HttpTimelineLoggingState> getHttpEnableTimelineLogging(
           String isolateId) async =>
       HttpTimelineLoggingState(enabled: httpEnableTimelineLoggingResult);
